Moving obsoletes back the TestableInvokeHandlerContext
